+++ app/base/config.js
@@ -56,7 +56,7 @@ angular.module('cmConfig',[])
                 'login': {
                     isDefault: true,
                     css: 'no-header no-footer',
-                    guests: false
+                    guests: true
                 },
                 'settings': {
                     routes: ['/settings'],
@@ -168,6 +168,23 @@ angular.module('cmConfig',[])
                     guests: true,
                     resolvePurl: true
                 },
+                'password-lost': {
+                    templateUrl: 'routes/password/lost/password-lost.html',
+                    routes:[
+                        '/password/lost',
+                        '/password'
+                    ],
+                    guests: true
+                },
+                'password-reset': {
+                    templateUrl: 'routes/password/reset/password-reset.html',
+                    routes:[
+                        '/password/reset/:resetId',
+                        '/passwordReset/:resetId'
+                    ],
+                    hasCtrl: true,
+                    guests: true
+                },
                 'registration': {
                     guests: true
                 },
@@ -210,7 +227,8 @@ angular.module('cmConfig',[])
                     hasCtrl: true
                 },
                 'verification': {
-                    routes:['/verification/:secret']
+                    routes:['/verification/:secret'],
+                    hasCtrl: true
                 },
                 'server_down' : {
                     templateUrl: 'routes/landingpages/server_down.html'