+++ core/comps/core/srvc-auth.js
@@ -10,7 +10,8 @@
  * @requires localStorage TODO: implement ServiceLocalStorage
  */
 
-angular.module('cmCore').service('cmAuth', [
+angular.module('cmCore')
+    .service('cmAuth', [
     'cmApi','LocalStorageAdapter', 'cmObject', 'cmUtil', 'cmLogger', 'cmCrypt' ,'$rootScope',
     function(cmApi, LocalStorageAdapter, cmObject, cmUtil, cmLogger, cmCrypt, $rootScope){
         var _TOKEN_ = undefined;
@@ -445,24 +446,43 @@ angular.module('cmCore').service('cmAuth', [
              */
             getTwoFactorToken: function(){
                 return localStorage.getItem('twoFactorToken');
+            },
+
+            sendPasswordLost: function(data){
+                return cmApi.post({
+                    path: '/resetPassword',
+                    data: data
+                });
+            },
+
+            resetPassword: function(data, resetId){
+                return cmApi.post({
+                    path: '/resetPassword/'+resetId,
+                    data: data
+                });
             }
         };
 
         cmObject.addEventHandlingTo(auth);
 
         cmApi.on('identity:update', function (event, data){
-//            cmLogger.debug('cmAuth.on:identity:update')
-            auth.trigger('identity:updated', data)
+            //cmLogger.debug('cmAuth.on:identity:update')
+            auth.trigger('identity:updated', data);
         });
 
         cmApi.on('identity:new', function (event, data){
             //cmLogger.debug('cmAuth.on:identity:new')
-            auth.trigger('identity:new', data)
+            auth.trigger('identity:new', data);
         });
 
         cmApi.on('conversation:new-aePassphrase', function(event, data){
-           //cmLogger.debug('conversation:new-aePassphrase');
-            auth.trigger('conversation:update', data)
+            //cmLogger.debug('cmAuth.on:conversation:new-aePassphrase');
+            auth.trigger('conversation:update', data);
+        });
+
+        cmApi.on('account:update', function (event, data){
+            //cmLogger.debug('cmAuth.on:account:update')
+            auth.trigger('account:update', data);
         });
 
         return auth;